Refactor to task class
